RGW - Move DAOS and MOTR to proper locations
